fix(security+backend+frontend): sessie 89 — multi-terminal audit fixes
Terminal A (security):
- Add auth to merge-fields and docx-templates endpoints
- RLS policy for email_logs table (new migration)
- Parameterize tenant SET LOCAL via set_config()
- Harden rate limiter X-Forwarded-For handling
- OAuth postMessage origin check
- Sanitizer: block img src tracker pixels, force noopener on links
- Add CSP upgrade-insecure-requests to Caddyfile

Terminal B (juridisch + backend):
- 14-dagenbrief: fix "na dagtekening" → "na ontvangst"
- Fix nakosten amounts in dutch-legal-rules.md
- Invoice numbering: FOR UPDATE to prevent race condition
- Float → Decimal in relations schemas
- invoice_parser: float(d) → str(d) for Decimal safety
- files_service: db.commit() → db.flush() (unit of work)
- Conflict check: add selectinload for client/opposing_party

Terminal C (frontend):
- Replace all window.confirm/prompt with ConfirmDialog component
- Add shadcn AlertDialog component
- Token refresh mutex to prevent race condition
- Remove duplicate formatDateShort from correspondentie
- Remove duplicate formatFileSize from hooks
- Centralize download helper in api.ts
- Clean up unused imports across settings tabs

+++ backend/app/cases/router.py
@@ -3,6 +3,7 @@
 import math
 import uuid
 from datetime import date
+from decimal import Decimal
 
 from fastapi import APIRouter, Depends, File, Form, HTTPException, Query, UploadFile, status
 from fastapi.responses import FileResponse, Response
@@ -45,8 +46,8 @@ async def list_cases(
     assigned_to_id: uuid.UUID | None = Query(default=None),
     date_from: date | None = Query(default=None),
     date_to: date | None = Query(default=None),
-    min_amount: float | None = Query(default=None, ge=0),
-    max_amount: float | None = Query(default=None, ge=0),
+    min_amount: Decimal | None = Query(default=None, ge=0),
+    max_amount: Decimal | None = Query(default=None, ge=0),
     is_active: bool = Query(default=True),
     current_user: User = Depends(get_current_user),
     db: AsyncSession = Depends(get_db),
